Refactored most of Style into StyleBase.
Ready for `ControlTheme` class, which is a style without a selector.
